Compilation fixes for MacOSX.
git-svn-id: http://svn.purei.org/purei/trunk@1170 b36208d7-6611-0410-8bec-b1987f11c4a2

+++ Source/GIF.cpp
@@ -244,7 +244,7 @@ uint32 CGIF::ProcessPacket(uint8* memory, uint32 address, uint32 end, const CGsP
 			{
 				if(tag.pre != 0)
 				{
-					writeList.push_back(CGSHandler::RegisterWrite(GS_REG_PRIM, tag.prim));
+					writeList.push_back(CGSHandler::RegisterWrite(GS_REG_PRIM, static_cast<uint64>(tag.prim)));
 				}
 			}
